Finish converting json-ld.js to TypeScript
