+++ common/model/breadcrumbs.ts
@@ -1,0 +1,6 @@
+export type Breadcrumbs = {
+  text: string;
+  url?: string;
+  prefix?: string;
+  isHidden?: boolean;
+}[];
